+++ packages/kity-runtime/src/boot/public-kity-src.d.ts
@@ -1,4 +1,0 @@
-declare module '../../../kity-assets/public/src/*.js' {
-  const value: any;
-  export default value;
-}