+++ packages/kity-runtime/src/create-editor.ts
@@ -1,4 +1,4 @@
-import { kityAssetManifest, type KityAssetManifest } from '@formulaxjs/kity-assets';
+import { kityAssetManifest, type KityAssetManifest } from './asset-manifest';
 import { legacyBoxType } from './vendor/legacy-box-type';
 import { legacyCharPosition } from './vendor/char-position';
 import { legacyOtherPosition } from './vendor/other-position';