+++ tsup.config.mjs
@@ -20,7 +20,6 @@ export default defineConfig({
     '.css': 'file',
   },
   alias: {
-    '@formulaxjs/kity-assets': path.resolve(__dirname, 'packages/kity-assets/src/index.ts'),
     '@formulaxjs/editor': path.resolve(__dirname, 'packages/editor/src/index.ts'),
     '@formulaxjs/renderer': path.resolve(__dirname, 'packages/renderer/src/index.ts'),
     '@formulaxjs/kity-runtime': path.resolve(__dirname, 'packages/kity-runtime/src/index.ts'),